🚑️ [긴급 수정] : 검색된 뉴스가 5개 이하일 시, 7일간의 뉴스 데이터를 Search하는 로직 추가
